style: phpcs-fix CompareRunnerTest and extract FakeCompareRunner
- Auto-fix 56 CS errors in tests/Unit/CompareRunnerTest.php via phpcbf
- Extract FakeCompareRunner to its own file to satisfy
  Squiz.Classes.ClassFileName.NoMatch and
  PSR1.Classes.ClassDeclaration.MultipleClasses
- Strip section-divider comments that tripped
  Squiz.Commenting.FunctionComment.WrongStyle
